lfs/manage/requirements_test.txt
# Core testing framework
pytest>=8.3.0
pytest-django>=4.10.0

# Test utilities  
pytest-mock>=3.14.0
pytest-cov>=6.2.0
pytest-xdist>=3.8.0
factory-boy>=3.3.0
coverage[toml]>=7.10.0
